+++ apps/meteor/client/startup/routes.tsx
@@ -6,7 +6,6 @@ import { router } from '../providers/RouterProvider';
 import MainLayout from '../views/root/MainLayout';
 
 const IndexRoute = lazy(() => import('../views/root/IndexRoute'));
-const MeetRoute = lazy(() => import('../views/meet/MeetRoute'));
 const HomePage = lazy(() => import('../views/home/HomePage'));
 const DirectoryPage = lazy(() => import('../views/directory'));
 const OmnichannelDirectoryRouter = lazy(() => import('../views/omnichannel/directory/OmnichannelDirectoryRouter'));
@@ -24,6 +23,7 @@ const ResetPasswordPage = lazy(() =>
 const OAuthAuthorizationPage = lazy(() => import('../views/oauth/OAuthAuthorizationPage'));
 const OAuthErrorPage = lazy(() => import('../views/oauth/OAuthErrorPage'));
 const NotFoundPage = lazy(() => import('../views/notFound/NotFoundPage'));
+const CallHistoryPage = lazy(() => import('../views/mediaCallHistory/CallHistoryPage'));
 
 declare module '@rocket.chat/ui-contexts' {
 	interface IRouterPaths {
@@ -107,6 +107,10 @@ declare module '@rocket.chat/ui-contexts' {
 			pathname: `/saml/${string}`;
 			pattern: '/saml/:token';
 		};
+		'call-history': {
+			pathname: `/call-history${`/details/${string}` | ''}`;
+			pattern: '/call-history/:tab?/:historyId?';
+		};
 	}
 }
 
@@ -126,11 +130,6 @@ router.defineRoutes([
 
 			return null;
 		}),
-	},
-	{
-		path: '/meet/:rid',
-		id: 'meet',
-		element: appLayout.wrap(<MeetRoute />),
 	},
 	{
 		path: '/home',
@@ -232,6 +231,15 @@ router.defineRoutes([
 		path: '/saml/:token',
 		id: 'saml',
 		element: appLayout.wrap(<SAMLLoginRoute />),
+	},
+	{
+		path: '/call-history/:tab?/:historyId?',
+		id: 'call-history',
+		element: appLayout.wrap(
+			<MainLayout>
+				<CallHistoryPage />
+			</MainLayout>,
+		),
 	},
 	{
 		path: '*',