+++ apps/meteor/client/views/omnichannel/directory/contacts/ContactTable.tsx
@@ -9,7 +9,6 @@ import {
 	usePagination,
 	useSort,
 } from '@rocket.chat/ui-client';
-import { useRoute } from '@rocket.chat/ui-contexts';
 import { hashKey } from '@tanstack/react-query';
 import { useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -18,15 +17,14 @@ import ContactTableRow from './ContactTableRow';
 import { useCurrentContacts } from './hooks/useCurrentContacts';
 import FilterByText from '../../../../components/FilterByText';
 import GenericNoResults from '../../../../components/GenericNoResults';
-import { useIsCallReady } from '../../../../contexts/CallContext';
 import { links } from '../../../../lib/links';
+import { useOmnichannelDirectoryRouter } from '../hooks/useOmnichannelDirectoryRouter';
 
 function ContactTable() {
 	const { t } = useTranslation();
 
 	const [term, setTerm] = useState('');
-	const directoryRoute = useRoute('omnichannel-directory');
-	const isCallReady = useIsCallReady();
+	const omnichannelDirectoryRouter = useOmnichannelDirectoryRouter();
 
 	const { current, itemsPerPage, setItemsPerPage, setCurrent, ...paginationProps } = usePagination();
 	const { sortBy, sortDirection, setSort } = useSort<'name' | 'channels.lastChat.ts' | 'contactManager.username' | 'lastChat.ts'>('name');
@@ -45,7 +43,7 @@ function ContactTable() {
 	);
 
 	const onButtonNewClick = useEffectEvent(() =>
-		directoryRoute.push({
+		omnichannelDirectoryRouter.navigate({
 			tab: 'contacts',
 			context: 'new',
 		}),
@@ -88,7 +86,6 @@ function ContactTable() {
 			>
 				{t('Last_Chat')}
 			</GenericTableHeaderCell>
-			{isCallReady && <GenericTableHeaderCell key='call' width={44} />}
 			<GenericTableHeaderCell key='spacer' w={40} />
 		</>
 	);