+++ apps/meteor/client/views/room/Header/HeaderIconWithRoom.tsx
@@ -18,4 +18,5 @@ const HeaderIconWithRoom = ({ room }: HeaderIconWithRoomProps): ReactElement => 
 
 	return <HeaderIcon icon={icon} />;
 };
+
 export default HeaderIconWithRoom;