+++ apps/meteor/client/views/room/Header/Omnichannel/QuickActions/QuickActions.tsx
@@ -32,7 +32,7 @@ const QuickActions = ({ className }: QuickActionsProps) => {
 				};
 
 				if (options) {
-					return <QuickActionOptions key={id} action={action} options={options} {...props} />;
+					return <QuickActionOptions options={options} key={id} action={action} {...props} />;
 				}
 
 				return <HeaderToolbarAction key={id} onClick={() => action(id)} {...props} />;