+++ apps/meteor/client/views/room/NotSubscribedRoom.tsx
@@ -1,7 +1,5 @@
 import type { IRoom } from '@rocket.chat/core-typings';
 import { Box, States, StatesAction, StatesActions, StatesIcon, StatesSubtitle, StatesTitle } from '@rocket.chat/fuselage';
-import { FeaturePreview, FeaturePreviewOff, FeaturePreviewOn, HeaderV1, HeaderV1Toolbar, SidebarToggler } from '@rocket.chat/ui-client';
-import { useLayout } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 
@@ -16,27 +14,10 @@ type NotSubscribedRoomProps = {
 
 const NotSubscribedRoom = ({ rid, reference, type }: NotSubscribedRoomProps): ReactElement => {
 	const { t } = useTranslation();
-
 	const handleJoinClick = useJoinRoom();
-
-	const { isMobile } = useLayout();
 
 	return (
 		<RoomLayout
-			header={
-				isMobile && (
-					<FeaturePreview feature='newNavigation'>
-						<FeaturePreviewOff>
-							<HeaderV1 justifyContent='start'>
-								<HeaderV1Toolbar>
-									<SidebarToggler />
-								</HeaderV1Toolbar>
-							</HeaderV1>
-						</FeaturePreviewOff>
-						<FeaturePreviewOn>{null}</FeaturePreviewOn>
-					</FeaturePreview>
-				)
-			}
 			body={
 				<Box display='flex' justifyContent='center' height='full'>
 					<States>