+++ apps/meteor/tests/e2e/page-objects/encrypted-room.ts
@@ -3,8 +3,8 @@ import { Message } from './fragments/message';
 import { DisableRoomEncryptionModal, EnableRoomEncryptionModal } from './fragments/modals';
 
 export class EncryptedRoomPage extends HomeContent {
-	get encryptedIcon() {
-		return this.page.locator('.rcx-room-header i.rcx-icon--name-key');
+	get encryptedTitle() {
+		return this.page.getByRole('button', { name: '- encrypted' });
 	}
 
 	get encryptionNotReadyIndicator() {